+++ src/canvas_sync/entities/linked_file.py
@@ -18,13 +18,12 @@ See developer_info.txt file for more information on the class hierarchy of Canva
 
 # Inbuilt modules
 import os
-import sys
 
 import requests
 
 # Canvas-Sync module imports
 from canvas_sync.entities.canvas_entity import CanvasEntity
-from canvas_sync.utilities.ANSI import ANSI
+from canvas_sync.utilities.console import console
 
 
 class LinkedFile(CanvasEntity):
@@ -69,8 +68,7 @@ class LinkedFile(CanvasEntity):
             " " * 15
             + "|   "
             + "\t" * self.indent
-            + "%s: %s"
-            % (ANSI.format("Linked File", formatting="linkedfile"), self.name)
+            + "[magenta]Linked File[/magenta]: %s" % self.name
         )
 
     def url_is_valid(self):
@@ -78,17 +76,15 @@ class LinkedFile(CanvasEntity):
 
     def print_status(self, status, color, overwrite_previous_line=False):
         """Print status to console"""
-
-        if overwrite_previous_line:
-            # Move up one line
-            sys.stdout.write(ANSI.format("", formatting="lineup"))
-            sys.stdout.flush()
-
-        print(
-            ANSI.format("[%s]" % status, formatting=color)
-            + str(self)[len(status) + 2 :]
-        )
-        sys.stdout.flush()
+        del overwrite_previous_line
+        style_map = {
+            "blue": "bold blue",
+            "green": "bold green",
+            "red": "bold red",
+            "yellow": "bold yellow",
+        }
+        style = style_map.get(color, "white")
+        console.print(f"[{style}][{status}][/{style}]{str(self)[len(status) + 2 :]}")
 
     def download(self):
         """
@@ -121,7 +117,7 @@ class LinkedFile(CanvasEntity):
 
     def walk(self, counter):
         """Stop walking, endpoint"""
-        print(str(self))
+        console.print(str(self))
 
         counter[0] += 1
         return
@@ -140,4 +136,4 @@ class LinkedFile(CanvasEntity):
 
     def show(self):
         """Show the folder hierarchy by printing every level"""
-        print(str(self))
+        console.print(str(self))