+++ src/canvas_sync/entities/page.py
@@ -21,13 +21,12 @@ See developer_info.txt file for more information on the class hierarchy of Canva
 import io
 import os
 import re
-import sys
 
 from canvas_sync.entities.canvas_entity import CanvasEntity
 from canvas_sync.entities.file import File
 from canvas_sync.entities.linked_file import LinkedFile
 from canvas_sync.utilities import helpers
-from canvas_sync.utilities.ANSI import ANSI
+from canvas_sync.utilities.console import console
 
 
 class Page(CanvasEntity):
@@ -73,7 +72,7 @@ class Page(CanvasEntity):
             " " * 15
             + "|   "
             + "\t" * self.indent
-            + "%s: %s" % (ANSI.format("Page", formatting="page"), self.name)
+            + "[cyan]Page[/cyan]: %s" % self.name
         )
 
     def download_linked_files(self, html_body):
@@ -159,20 +158,19 @@ class Page(CanvasEntity):
 
     def print_status(self, status, color, overwrite_previous_line=False):
         """Print status to console"""
-        if overwrite_previous_line:
-            # Move up one line
-            sys.stdout.write(ANSI.format("", formatting="lineup"))
-            sys.stdout.flush()
-
-        print(
-            ANSI.format("[%s]" % status, formatting=color)
-            + str(self)[len(status) + 2 :]
-        )
-        sys.stdout.flush()
+        del overwrite_previous_line
+        style_map = {
+            "blue": "bold blue",
+            "green": "bold green",
+            "red": "bold red",
+            "yellow": "bold yellow",
+        }
+        style = style_map.get(color, "white")
+        console.print(f"[{style}][{status}][/{style}]{str(self)[len(status) + 2 :]}")
 
     def walk(self, counter):
         """Stop walking, endpoint"""
-        print(str(self))
+        console.print(str(self))
 
         counter[0] += 1
         return
@@ -195,4 +193,4 @@ class Page(CanvasEntity):
 
     def show(self):
         """Show the folder hierarchy by printing every level"""
-        print(str(self))
+        console.print(str(self))